requirements.txt
openpyxl>=3.1.2
jinja2>=3.1.3
requests>=2.31.0
google-generativeai>=0.4.0
